chore: refactor to allow new strategy structure
