+++ lib/src/main/java/nl/dflipse/fit/strategy/Generator.java
@@ -1,0 +1,9 @@
+package nl.dflipse.fit.strategy;
+
+import java.util.List;
+
+import nl.dflipse.fit.faultload.Fault;
+
+public interface Generator {
+    public List<Fault> generate(FaultloadResult result);
+}
