+++ lib/src/main/java/nl/dflipse/fit/strategy/PruningStrategy.java
@@ -1,0 +1,9 @@
+package nl.dflipse.fit.strategy;
+
+import java.util.List;
+
+import nl.dflipse.fit.faultload.Faultload;
+
+public interface PruningStrategy {
+    public boolean filter(Faultload faultload, List<FaultloadResult> historicResults);
+}
